# PERF: groupby.pad/backfill follow-up