- [ ] closes #xxxx
- [ ] tests added / passed
- [ ] Ensure all linting tests pass, see [here](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit) for how to run them
- [ ] whatsnew entry

Follow-ups include the big perf bump discussed in #43478

```
import pandas as pd
import numpy as np

np.random.seed(23446365)
arr = np.random.randn(10**5, 10)
mask = arr < -1
arr[mask] = np.nan

df = pd.DataFrame(arr)

gb = df.groupby(df.index % 7)

%timeit res = gb.pad()
29.8 ms ± 193 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)  # <- master
14.6 ms ± 2.41 ms per loop (mean ± std. dev. of 7 runs, 100 loops each)  # <- PR
```