@jreback 2021-09-12
prob should do some kind of whatsnew note for significant improvements like this (followon)